Make roland run as a unique application.
This basically means BrowserWindow is now a Gtk.Window, and Roland is a
Gtk.Application. Download management was the difficult part, as we need
to ensure at least one window is alive.

This also fixes the situation where you have a single window open with
downloads going, and you close it. Now it will prompt about cancelling
in-progress downloads.