Port to WebKit2.
There are bound to be a lot of things broken with this, but for the most
part this is now working.

WebKit2 is the future, it uses less memory, and has more features that I
care about.

This also fixes a few bugs in download management, tab completion, as
well as adding a command to dump certificate information for the current
domain.